move encrypt_and_execute_query to ipa-core/tests/encrypted_input
