+++ ipa-core/src/cli/mod.rs
@@ -1,11 +1,6 @@
 #[cfg(feature = "web-app")]
 mod clientconf;
-#[cfg(all(
-    feature = "test-fixture",
-    feature = "web-app",
-    feature = "cli",
-    feature = "in-memory-infra"
-))]
+#[cfg(all(feature = "test-fixture", feature = "web-app", feature = "cli"))]
 pub mod crypto;
 mod csv;
 mod ipa_output;
